feat: refresh group page navigation (#15)
