+++ web/src/App.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { Layout } from './components/Layout'
 import HomePage from './pages/HomePage'
 import GroupPage from './pages/GroupPage'
+import GroupDashboardPage from './pages/GroupDashboardPage'
 import EventsListPage from './pages/EventsListPage'
 import NewEventPage from './pages/NewEventPage'
 import EventDetailPage from './pages/EventDetailPage'
@@ -14,7 +15,8 @@ export default function App() {
       <Routes>
         <Route element={<Layout />}>
           <Route index element={<HomePage />} />
-          <Route path="groups/:groupId" element={<GroupPage />}>            
+          <Route path="groups/:groupId" element={<GroupPage />}>
+            <Route index element={<GroupDashboardPage />} />
             <Route path="events" element={<EventsListPage />} />
             <Route path="events/new" element={<NewEventPage />} />
             <Route path="events/:eventId" element={<EventDetailPage />} />
